move restricted to commands.decorators
