+++ commands/decorators/__init__.py
@@ -1,2 +1,3 @@
 from .log import log
+from .restricted import restricted
 from .send_typing_action import send_typing_action
