+++ commands/decorators/restricted.py
@@ -1,0 +1,17 @@
+import logging
+from functools import wraps
+
+from secrets import LIST_OF_ADMINS
+
+
+def restricted(func):
+    '''decorator that restricts use to only the admins listed in secrets.py'''
+    @wraps(func)
+    def wrapped(update, context, *args, **kwargs):
+        user_id = update.effective_user.id
+        name = update.effective_user.username
+        if user_id not in LIST_OF_ADMINS:
+            logging.info(f'Unauthorized access: {name} ({user_id}) tried to use {func.__name__}()')
+            return
+        return func(update, context, *args, **kwargs)
+    return wrapped
